refactor flush-cache related code
