+++ src/arch/loongarch64/zone.rs
@@ -690,4 +690,8 @@ impl Zone {
     pub fn arch_zone_post_configuration(&mut self, config: &HvZoneConfig) -> HvResult {
         Ok(())
     }
+
+    pub fn arch_zone_reset(&mut self, config: &HvZoneConfig) -> HvResult {
+        Ok(())
+    }
 }
