+++ src/arch/x86_64/zone.rs
@@ -149,6 +149,10 @@ impl Zone {
 
         Ok(())
     }
+
+    pub fn arch_zone_reset(&mut self, config: &HvZoneConfig) -> HvResult {
+        Ok(())
+    }
 }
 
 /*impl BarRegion {
